requirements.txt
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-WTF==1.1.1
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
numpy==1.26.0
pandas==2.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
six==1.16.0
tzdata==2023.3
Werkzeug==2.3.7
WTForm==1.0
WTForms==3.0.1
